use diagnostic colors
Summary:
This is supported in gcc 4.9.y and greater and only works if output is a terminal via a gcc check, so as not to break other types of builds. This is not intended to colorize everything, only errors, warnings, notes, carets, locus', or quotes. Documentation for this feature may be found here: http://gcc.gnu.org/onlinedocs/gcc/Language-Independent-Options.html
It is worth noting that the documentation is slightly outdated, due to the cflag being removed in favor of setting the equivalent to -fdiagnostics-color=always anytime GCC_COLORS is detected in the environment.

Test Plan: Fuck test plans, this is the dopest dope you've ever smoked

Reviewers: randall.rushing, Mustaavalkosta, michyprima, rlasalle15

Differential Revision: http://review.codefirex.com/D59